refactor: moved game-zone images in folders
