+++ src/assets/img/game-zone/currency/pot-amount.png
