+++ src/assets/img/outcomes/you-fold.png
