+++ src/assets/img/outcomes/you-lose.png
