+++ src/assets/styles/game-zone/game-zone.scss
@@ -493,7 +493,7 @@
   align-items: center;
 }
 
-.shadowquestion {
+.mystery-card {
   width: 1000px;
   height: 250px;
 }
