+++ src/components/game-table/PlayingCard.vue
@@ -19,7 +19,7 @@
 
 <script setup lang="ts">
 import { computed, watchEffect } from 'vue'
-import BackCard from '@/assets/img/back-card.png'
+import BackCard from '@/assets/img/cards/special-cards/back-card.png'
 import { useDataFetcher } from '@/composables/useDataFetcher'
 import type { Card } from '@/interface/card'
