+++ src/utils/gameplay/pop-ups/modalUtil.ts
@@ -1,8 +1,8 @@
 import { ref } from 'vue'
 import { useGameStore } from '@/stores/game-store'
-import YouWinImage from '@/assets/img/game-zone/you-win.png'
-import YouLoseImage from '@/assets/img/game-zone/you-lose.png'
-import YouFoldImage from '@/assets/img/game-zone/you-fold.png'
+import YouWinImage from '@/assets/img/outcomes/you-win.png'
+import YouLoseImage from '@/assets/img/outcomes/you-lose.png'
+import YouFoldImage from '@/assets/img/outcomes/you-fold.png'
 import YouWinSound from '@/assets/sfx/voices/you-win.wav'
 import YouLoseSound from '@/assets/sfx/voices/you-lose.wav'
 import YouFoldSound from '@/assets/sfx/outcomes/lose_1.wav'
